Notices refactoring
