+++ view/adminhtml/templates/common.phtml
@@ -5,8 +5,6 @@
 /** @var \Algolia\AlgoliaSearch\ViewModel\Adminhtml\Common $viewModel */
 $viewModel = $this->getViewModel();
 
-$applicationId = $viewModel->getApplicationId();
-
 $isClickAnalyticsEnabled = true;
 $isClickAnalyticsTurnedOnInAdmin = true;
 $isQueryRulesEnabled = true;
@@ -22,10 +20,6 @@ if ($section === 'algoliasearch_instant') {
     $isQueryRulesEnabled = $viewModel->isQueryRulesEnabled();
 }
 
-$indexingQueueInfoUrl = $this->getUrl('algolia_algoliasearch/queue/info');
-$indexingQueueConfigUrl = $this->getUrl('adminhtml/system_config/edit/section/algoliasearch_queue');
-$indexingQueuePageUrl = $this->getUrl('algolia_algoliasearch/queue/index');
-
 $ajaxCheckUrl = $this->getUrl('algolia_algoliasearch/landingpage/checkUrl');
 $ajaxCheckQuery = $this->getUrl('algolia_algoliasearch/query/checkQuery');
 
@@ -38,28 +32,25 @@ $icons = [
 $videoConfig = $viewModel->getVideoConfig($section);
 $linksConfig = $viewModel->getLinksConfig($section);
 
-$isMsiExternalModuleNeeded = $viewModel->isMsiExternalModuleNeeded();
+
 $isEsWarningNeeded = $viewModel->isEsWarningNeeded();
-$newVersionNotification = $viewModel->getNewVersionNotification();
+$extensionNotices = $viewModel->getExtensionNotices();
 
 ?>
 
 <script>
-var applicationId = <?php echo json_encode($applicationId); ?>;
 var isClickAnalyticsEnabled = <?php echo json_encode($isClickAnalyticsEnabled); ?>;
 var isClickAnalyticsTurnedOnInAdmin = <?php echo json_encode($isClickAnalyticsTurnedOnInAdmin); ?>;
 var isQueryRulesEnabled = <?php echo json_encode($isQueryRulesEnabled); ?>;
-var indexingQueueConfigUrl = <?php echo json_encode($indexingQueueConfigUrl); ?>;
-var indexingQueuePageUrl = <?php echo json_encode($indexingQueuePageUrl); ?>;
-var indexingQueueInfoUrl = <?php echo json_encode($indexingQueueInfoUrl); ?>;
+
 var ajaxCheckUrl = <?php echo json_encode($ajaxCheckUrl); ?>;
 var ajaxCheckQuery = <?php echo json_encode($ajaxCheckQuery); ?>;
 var icons = <?php echo json_encode($icons); ?>;
 var videoConfig = <?php echo json_encode($videoConfig); ?>;
 var linksConfig = <?php echo json_encode($linksConfig); ?>;
-var isMsiExternalModuleNeeded = <?php echo json_encode($isMsiExternalModuleNeeded); ?>;
+
 var isEsWarningNeeded = <?php echo json_encode($isEsWarningNeeded); ?>;
-var newVersionNotification = <?php echo json_encode($newVersionNotification); ?>;
+var extensionNotices = <?php echo json_encode($extensionNotices); ?>;
 
 document.addEventListener("DOMContentLoaded", function(event) {
 	requirejs(['algoliaAdminBundle'], function(algoliaAdminBundle) {
@@ -137,65 +128,8 @@ document.addEventListener("DOMContentLoaded", function(event) {
 				$('#row_algoliasearch_instant_instant_facets .algolia_block.blue').after(notEnabledText);
 			}
 
-			// QUEUE INFO
-
 			var url = window.location.href;
 			if (url.indexOf('section/algoliasearch_') !== -1) {
-				$.getJSON(indexingQueueInfoUrl, function(queueInfo) {
-					var message = `
-						<div class="algolia_block icon-warning">
-							<div class="heading"><a href="` + indexingQueueConfigUrl + `">Indexing Queue</a> is not enabled</div>
-							It is highly recommended that you enable it, especially if you are on a production environment.
-							<br><br>
-							Find out more about Indexing Queue in <a href="https://www.algolia.com/doc/integration/magento-2/how-it-works/indexing-queue/?utm_source=magento&utm_medium=extension&utm_campaign=magento_2&utm_term=shop-owner&utm_content=doc-link" target="_blank">documentation</a>.
-						</div>`;
-
-					if (queueInfo.isEnabled === true) {
-						message = `
-							<div class="algolia_block icon-bulb">
-								<div class="heading">Queued indexing jobs</div>
-								Number of queued jobs: <strong>` + queueInfo.currentSize + `</strong>.
-								Assuming your queue runner runs every 5 minutes, all jobs will be processed
-								in approx. ` + queueInfo.eta + `.
-								You may want to <a href="` + indexingQueuePageUrl + `">clear the queue</a> or <a href="` + indexingQueueConfigUrl + `">configure indexing queue</a>.
-								<br><br>
-								Find out more about Indexing Queue in <a href="https://www.algolia.com/doc/integration/magento-2/how-it-works/indexing-queue/?utm_source=magento&utm_medium=extension&utm_campaign=magento_2&utm_term=shop-owner&utm_content=doc-link" target="_blank">documentation</a>.
-							</div>`;
-					}
-
-					$('.entry-edit').before(message);
-
-					if (isMsiExternalModuleNeeded === true) {
-						var msiMessage = `
-						<div class="algolia_block icon-warning">
-							<div class="heading">Magento Multi-source Inventory compatibility</div>
-							Your store is using Magento Multi-source Inventory and for Algolia to be fully compatible with it, you should install the module <a target="_blank" href="https://github.com/algolia/algoliasearch-inventory-magento-2/">Algoliasearch Inventory</a>.<br>
-							More information in <a href="https://www.algolia.com/doc/integration/magento-2/guides/multi-source-inventory/?utm_source=magento&utm_medium=extension&utm_campaign=magento_2&utm_term=shop-owner&utm_content=doc-link" target="_blank">documentation</a>.
-							<br>
-						</div>`;
-						$('.entry-edit').before(msiMessage);
-					}
-				});
-				// New version notification
-				if (newVersionNotification) {
-					notification = `
-						<div class="algolia_block icon-warning">
-							<div class="heading">Algolia Extension update</div>
-							You are using old version of Algolia extension. Latest version of the extension is v <b>` + newVersionNotification.version + `</b><br />
-							It is highly recommended to update your version to avoid any unexpecting issues and to get new features.<br />
-							See details on our <a target="_blank" href="` + newVersionNotification.url + `">Github repository</a>.
-						</div>`;
-					$('.entry-edit').before(notification);
-				}
-
-				// ES compatibility warning
-				if (isEsWarningNeeded) {
-					var esMessage = `<span class="algolia-config-warning">&#9888;</span>This feature is currently  supported only for MySQL search engine.`;
-				} else {
-					var esMessage = `<span class="algolia-config-warning">&#9888;</span> If you turn on this feature, please make sure that every attribute you choose in the facet listing below is configured as "<strong>Use in Layered Navigation</strong>" for categories and "<strong>Use in Search Result Layered Navigation</strong>" for Catalog Search in <strong>Stores > Attributes > Product > "Storefront Properties" panel of attribute configuration</strong>.
-						`;
-				}
-				$('#row_algoliasearch_instant_instant_backend_rendering_enable .note').append(esMessage);
 
 				// Admin videos
                 var videoTemplate = '';
@@ -232,6 +166,17 @@ document.addEventListener("DOMContentLoaded", function(event) {
                     <hr>
                     <br>`;
                 $('.algolia-admin-content').html(linksAndVideoTemplate);
+
+
+            	// New notice management
+            	extensionNotices.each(function(notice){
+            		if (notice.method === 'before') {
+        				$(notice.selector).before(notice.message);
+        			} else if (notice.method === 'append') {
+						$(notice.selector).append(notice.message);
+        			}
+            	});
+
 			}
 
 
@@ -285,7 +230,7 @@ document.addEventListener("DOMContentLoaded", function(event) {
 
 			// EVENT TRACKING
             algoliaEventsTracking.track();
-
+            
 		});
 	});
